merge: :twisted_rightwards_arrows: merge abstractions to staging
