+++ bot/index.js
@@ -46,6 +46,16 @@ import {
   applyCodemetaMetadata,
   applyCitationMetadata,
 } from "./metadata/index.js";
+import {
+  publishToZenodo,
+  reRenderDashboard,
+} from "./commands/actions/index.js";
+import {
+  rerunCWLValidation,
+  rerunFullRepoValidation,
+  rerunLicenseValidation,
+  rerunMetadataValidation,
+} from "./commands/validations/index.js";
 
 checkEnvVariable("GH_APP_NAME");
 checkEnvVariable("CODEFAIR_APP_DOMAIN");
@@ -450,685 +460,31 @@ export default async (app, { getRouter }) => {
 
     // "API" using comments to trigger workflows
     if (issueBody.includes("<!-- @codefair-bot rerun-cwl-validation -->")) {
-      try {
-        logwatch.start("Rerunning CWL Validation...");
-
-        const [licenseResponse, metadataResponse, cwlResponse] =
-          await Promise.all([
-            db.licenseRequest.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-            db.codeMetadata.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-            db.cwlValidation.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-          ]);
-
-        const license = !!licenseResponse?.license_id;
-        const citation = !!metadataResponse?.contains_citation;
-        const codemeta = !!metadataResponse?.contains_codemeta;
-
-        const cwlObject = await getCWLFiles(context, owner, repository);
-
-        logwatch.info("at here!");
-
-        const subjects = {
-          cwl: cwlObject,
-          citation,
-          codemeta,
-          license,
-        };
-
-        const issueBody = await renderIssues(
-          context,
-          owner,
-          repository,
-          false,
-          subjects
-        );
-
-        logwatch.info("at heredfasdfad!");
-        await createIssue(context, owner, repository, ISSUE_TITLE, issueBody);
-
-        logwatch.info("CWL Validation rerun successfully!");
-      } catch (error) {
-        // Remove the command from the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-        if (error.cause) {
-          logwatch.error(
-            {
-              message: "Error.cause message for CWL Validation",
-              error_cause: error.cause,
-              error: error,
-            },
-            true
-          );
-        }
-        throw new Error("Error rerunning cwl validation", error);
-      }
+      await rerunCWLValidation(context, owner, repository, issueBody);
     }
 
     if (
       issueBody.includes("<!-- @codefair-bot rerun-full-repo-validation -->")
     ) {
-      logwatch.start("Rerunning full repository validation...");
-      try {
-        let subjects = await checkForCompliance(
-          context,
-          owner,
-          repository.name
-        );
-
-        // If existing cwl validation exists, update the contains_cwl value
-        const cwlExists = await db.cwlValidation.findUnique({
-          where: {
-            repository_id: repository.id,
-          },
-        });
-
-        if (cwlExists?.contains_cwl_files) {
-          cwlObject.contains_cwl_files = cwlExists.contains_cwl_files;
-
-          if (cwlExists.files.length > 0) {
-            // Remove the files that are not in cwlObject
-            const cwlFilePaths = cwlObject.files.map((file) => file.path);
-            cwlObject.removed_files = cwlExists.files.filter((file) => {
-              return !cwlFilePaths.includes(file.path);
-            });
-          }
-        }
-
-        subjects.cwl = cwlObject;
-
-        const issueBody = await renderIssues(
-          context,
-          owner,
-          repository,
-          false,
-          subjects
-        );
-
-        await createIssue(context, owner, repository, ISSUE_TITLE, issueBody);
-      } catch (error) {
-        // Remove the command from the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-        if (error.cause) {
-          logwatch.error(
-            {
-              message: "Error.cause message for Full Repo Validation",
-              error: error.cause,
-            },
-            true
-          );
-        }
-        throw new Error("Error rerunning full repo validation", error);
-      }
+      await rerunFullRepoValidation(context, owner, repository, issueBody);
     }
 
     if (issueBody.includes("<!-- @codefair-bot rerun-license-validation -->")) {
-      // Run the license validation again
-      logwatch.start("Rerunning License Validation...");
-      try {
-        const licenseRequest = await context.octokit.rest.licenses.getForRepo({
-          owner,
-          repo: repository.name,
-        });
-
-        const existingLicense = await db.licenseRequest.findUnique({
-          where: {
-            repository_id: repository.id,
-          },
-        });
-
-        const license = !!licenseRequest.data.license;
-
-        if (!license) {
-          throw new Error("License not found in the repository");
-        }
-
-        const { licenseId, licenseContent, licenseContentEmpty } =
-          validateLicense(licenseRequest, existingLicense);
-
-        logwatch.info({
-          message: `License validation complete`,
-          licenseId,
-          licenseContent,
-          licenseContentEmpty,
-        });
-
-        // Update the database with the license information
-        if (existingLicense) {
-          await db.licenseRequest.update({
-            data: {
-              license_id: licenseId,
-              license_content: licenseContent,
-              license_status: licenseContentEmpty ? "invalid" : "valid",
-            },
-            where: {
-              repository_id: repository.id,
-            },
-          });
-        } else {
-          await db.licenseRequest.create({
-            data: {
-              license_id: licenseId,
-              license_content: licenseContent,
-              license_status: licenseContentEmpty ? "valid" : "invalid",
-            },
-            where: {
-              repository_id: repository.id,
-            },
-          });
-        }
-
-        // Update the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-      } catch (error) {
-        // Remove the command from the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-        if (error.cause) {
-          logwatch.error(
-            {
-              message: "Error.cause message for License Validation",
-              error: error.cause,
-            },
-            true
-          );
-        }
-        throw new Error("Error rerunning license validation", error);
-      }
+      await rerunLicenseValidation(context, owner, repository, issueBody);
     }
 
     if (
       issueBody.includes("<!-- @codefair-bot rerun-metadata-validation -->")
     ) {
-      logwatch.start("Validating metadata files...");
-      try {
-        let metadata = await gatherMetadata(context, owner, repository);
-        let containsCitation = false,
-          containsCodemeta = false,
-          validCitation = false,
-          validCodemeta = false;
-
-        const existingMetadataEntry = await db.codeMetadata.findUnique({
-          where: {
-            repository_id: repository.id,
-          },
-        });
-
-        if (existingMetadataEntry?.metadata) {
-          // Update the metadata variable
-          containsCitation = existingMetadataEntry.contains_citation;
-          containsCodemeta = existingMetadataEntry.contains_codemeta;
-          metadata = applyDbMetadata(existingMetadataEntry, metadata);
-        }
-
-        const citation = await getCitationContent(context, owner, repository);
-        const codemeta = await getCodemetaContent(context, owner, repository);
-
-        if (codemeta) {
-          containsCodemeta = true;
-          validCodemeta = await validateMetadata(
-            codemeta,
-            "codemeta",
-            repository
-          );
-          metadata = await applyCodemetaMetadata(
-            codemeta,
-            metadata,
-            repository
-          );
-        }
-
-        if (citation) {
-          containsCitation = true;
-          validCitation = await validateMetadata(
-            citation,
-            "citation",
-            repository
-          );
-          metadata = await applyCitationMetadata(
-            citation,
-            metadata,
-            repository
-          );
-          // consola.info("Metadata so far after citation update", JSON.stringify(metadata, null, 2));
-        }
-
-        // Ensure all dates have been converted to ISO strings split by the T
-        if (metadata.creationDate) {
-          metadata.creationDate = convertDateToUnix(metadata.creationDate);
-        }
-        if (metadata.firstReleaseDate) {
-          metadata.firstReleaseDate = convertDateToUnix(
-            metadata.firstReleaseDate
-          );
-        }
-        if (metadata.currentVersionReleaseDate) {
-          metadata.currentVersionReleaseDate = convertDateToUnix(
-            metadata.currentVersionReleaseDate
-          );
-        }
-
-        // update the database with the metadata information
-        if (existingMetadataEntry) {
-          await db.codeMetadata.update({
-            data: {
-              codemeta_status: validCodemeta ? "valid" : "invalid",
-              citation_status: validCitation ? "valid" : "invalid",
-              contains_citation: containsCitation,
-              contains_codemeta: containsCodemeta,
-              metadata: metadata,
-            },
-            where: {
-              repository_id: repository.id,
-            },
-          });
-        } else {
-          await db.codeMetadata.create({
-            data: {
-              codemeta_status: validCodemeta ? "valid" : "invalid",
-              citation_status: validCitation ? "valid" : "invalid",
-              contains_citation: containsCitation,
-              contains_codemeta: containsCodemeta,
-              metadata: metadata,
-            },
-            where: {
-              repository_id: repository.id,
-            },
-          });
-        }
-
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-      } catch (error) {
-        // Remove the command from the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-        if (error.cause) {
-          logwatch.error(
-            {
-              message: "Error.cause message for Metadata Validation",
-              error: error.cause,
-            },
-            true
-          );
-        }
-        throw new Error("Error rerunning metadata validation", error);
-      }
+      await rerunMetadataValidation(context, owner, repository, issueBody);
     }
 
     if (issueBody.includes("<!-- @codefair-bot publish-zenodo")) {
-      logwatch.start("Publishing to Zenodo...");
-      const issueBodyRemovedCommand = issueBody.substring(
-        0,
-        issueBody.indexOf("<!-- @codefair-bot publish-zenodo")
-      );
-      const issueBodyNoArchiveSection = issueBodyRemovedCommand.substring(
-        0,
-        issueBody.indexOf("## FAIR Software Release")
-      );
-      const badgeURL = `${CODEFAIR_DOMAIN}/dashboard/${owner}/${repository.name}/release/zenodo`;
-      const releaseBadge = `[![Create Release](https://img.shields.io/badge/Create_Release-00bcd4.svg)](${badgeURL})`;
-      const { depositionId, releaseId, tagVersion, userWhoSubmitted } =
-        parseZenodoInfo(issueBody);
-      logwatch.info(
-        {
-          message: `Parsed Zenodo info:`,
-          depositionId,
-          releaseId,
-          tagVersion,
-          userWhoSubmitted,
-        },
-        true
-      );
-
-      try {
-        // 1. Get the metadata from the repository
-        const citationCff = await getCitationContent(
-          context,
-          owner,
-          repository
-        );
-        const codemeta = await getCodemetaContent(context, owner, repository);
-
-        // 2. Validate the CITATION.cff and codemeta.json files
-        await validateMetadata(citationCff, "citation", repository);
-        await validateMetadata(codemeta, "codemeta", repository);
-
-        // 3. Fetch the Zenodo token from the database and verify it is valid
-        const zenodoToken = await getZenodoToken(userWhoSubmitted);
-
-        // 4. Create the Zenodo record or get the existing one and create a new draft deposition if none exist
-        const zenodoDepositionInfo = await getZenodoDepositionInfo(
-          depositionId,
-          zenodoToken
-        );
-
-        // Sending the Zenodo deposition info to the log in the case of an error
-        logwatch.info(
-          {
-            message: "Zenodo Deposition Info: ",
-            zenodoDepositionInfo,
-          },
-          true
-        );
-
-        // 4.5 Set the bucket URL and DOI
-        // using record_id over id because it the id of the current deposition
-        const newDepositionId = zenodoDepositionInfo.record_id;
-        const bucket_url = zenodoDepositionInfo.links.bucket;
-        const zenodoDoi = zenodoDepositionInfo.metadata.prereserve_doi.doi;
-
-        // Update progress in the GitHub issue
-        const tempString = `${issueBodyNoArchiveSection}\n\n## FAIR Software Release 🔄\n***${tagVersion}*** of your software is being released on GitHub and archived on Zenodo. A draft deposition was created and will be adding the necessary files and metadata.`;
-        const finalTempString = await applyLastModifiedTemplate(tempString);
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          finalTempString
-        );
-
-        // 5. Update the CITATION.cff and codemeta.json files with the DOI provided by Zenodo
-        const updatedMetadataFile = await updateMetadataIdentifier(
-          context,
-          owner,
-          repository,
-          zenodoDoi,
-          tagVersion
-        );
-
-        // 6. Gather metadata for Zenodo deposition
-        const newZenodoMetadata = await createZenodoMetadata(
-          updatedMetadataFile,
-          repository
-        );
-
-        // 7. Update the Zenodo deposition's metadata
-        await updateZenodoMetadata(
-          newDepositionId,
-          zenodoToken,
-          newZenodoMetadata
-        );
-
-        // 7.5 Get the GitHub draft release from the repository
-        const draftRelease = await getReleaseById(
-          context,
-          repository.name,
-          owner,
-          releaseId
-        );
-
-        const repositoryArchive = await downloadRepositoryZip(
-          context,
-          owner,
-          repository.name
-        );
-
-        await uploadReleaseAssetsToZenodo(
-          zenodoToken,
-          draftRelease.data.assets,
-          repositoryArchive,
-          owner,
-          context,
-          bucket_url,
-          repository,
-          tagVersion
-        );
-
-        // Update the GitHub issue with a status report
-        const afterUploadString = `${issueBodyNoArchiveSection}\n\n## FAIR Software Release 🔄\n***${tagVersion}*** of your software is being released on GitHub and archived on Zenodo. All assets from the GitHub repository's draft release have been successfully uploaded to the Zenodo deposition draft.`;
-        const finalUploadString =
-          await applyLastModifiedTemplate(afterUploadString);
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          finalUploadString
-        );
-
-        // 8. Publish the Zenodo deposition
-        await publishZenodoDeposition(zenodoToken, newDepositionId);
-
-        // Update the release to not be a draft
-        await updateGitHubRelease(context, repository.name, owner, releaseId);
-
-        // 9. Append to the issueBody that the deposition has been published
-        // Update the issue with the new body
-        const badge = `[![DOI](https://img.shields.io/badge/DOI-${zenodoDoi}-blue)](${ZENODO_ENDPOINT}/records/${newDepositionId})`;
-        const issueBodyArchiveSection = `${issueBodyNoArchiveSection}\n\n## FAIR Software Release ✔️\n***${tagVersion}*** of your software was successfully released on GitHub and archived on Zenodo. You can view the Zenodo archive by clicking the button below:\n\n${badge}\n\nReady to create your next FAIR release? Click the button below:\n\n${releaseBadge}`;
-        const finalTemplate = await applyLastModifiedTemplate(
-          issueBodyArchiveSection
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          finalTemplate
-        );
-
-        // Update the database with the Zenodo ID and the status
-        await db.zenodoDeposition.update({
-          data: {
-            status: "published",
-            zenodo_id: newDepositionId,
-            existing_zenodo_deposition_id: true,
-            last_published_zenodo_doi: zenodoDoi,
-          },
-          where: {
-            repository_id: repository.id,
-          },
-        });
-
-        logwatch.info("Updated the Zenodo deposition in the database!");
-
-        await db.analytics.update({
-          data: {
-            zenodo_release: {
-              increment: 1,
-            },
-            create_release: {
-              increment: 1,
-            },
-          },
-          where: {
-            id: repository.id,
-          },
-        });
-
-        logwatch.success("Updated the analytics in the database!");
-      } catch (error) {
-        // Update the issue with the new body
-        // Update the GitHub issue with a status report
-        const afterUploadString = `${issueBodyNoArchiveSection}\n\n## FAIR Software Release ❌\n***${tagVersion}*** of your software was not successfully released on GitHub and archived on Zenodo. There was an error during the publication process. Please try again later or reach out to the Codefair team for additional help.`;
-        const finalUploadString =
-          await applyLastModifiedTemplate(afterUploadString);
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          finalUploadString
-        );
-        await db.zenodoDeposition.update({
-          data: {
-            status: "error",
-          },
-          where: {
-            repository_id: repository.id,
-          },
-        });
-        if (error.cause) {
-          logwatch.error(
-            {
-              message: "Error.cause message for Zenodo Publishing",
-              error: error.cause,
-            },
-            true
-          );
-        }
-        throw new Error(`Error publishing to Zenodo: ${error.message}`, {
-          cause: error,
-        });
-      }
+      await publishToZenodo(context, owner, repository, issueBody);
     }
 
     if (issueBody.includes("<!-- @codefair-bot re-render-dashboard -->")) {
-      // Run database queries in parallel using Promise.all
-      logwatch.start("Re-rendering issue dashboard...");
-      try {
-        const [licenseResponse, metadataResponse, cwlResponse] =
-          await Promise.all([
-            db.licenseRequest.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-            db.codeMetadata.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-            db.cwlValidation.findUnique({
-              where: {
-                repository_id: repository.id,
-              },
-            }),
-          ]);
-
-        const license = !!licenseResponse?.license_id;
-        const citation = !!metadataResponse?.contains_citation;
-        const codemeta = !!metadataResponse?.contains_codemeta;
-        const cwl = !!cwlResponse?.contains_cwl_files;
-
-        const cwlObject = {
-          contains_cwl_files: cwl,
-          files: cwlResponse?.files || [],
-          removed_files: [],
-        };
-
-        const subjects = {
-          citation,
-          codemeta,
-          cwl: cwlObject,
-          license,
-        };
-
-        const issueBody = await renderIssues(
-          context,
-          owner,
-          repository,
-          false,
-          subjects
-        );
-
-        await createIssue(context, owner, repository, ISSUE_TITLE, issueBody);
-      } catch (error) {
-        // Remove the command from the issue body
-        const issueBodyRemovedCommand = issueBody.substring(
-          0,
-          issueBody.indexOf(`<sub><span style="color: grey;">Last updated`)
-        );
-        const lastModified = await applyLastModifiedTemplate(
-          issueBodyRemovedCommand
-        );
-        await createIssue(
-          context,
-          owner,
-          repository,
-          ISSUE_TITLE,
-          lastModified
-        );
-        throw new Error("Error rerunning re-rendering dashboard", error);
-      }
+      await reRenderDashboard(context, owner, repository, issueBody);
     }
   });
